requirements.txt
appdirs==1.4.3
argcomplete==1.10.0
argh==0.26.1
asn1crypto==0.24.0
attrs==19.3.0
Babel==2.7.0
backcall==0.1.0
bcrypt==3.1.6
Beaker==1.10.0
beautifulsoup4==4.9.1
bleach==3.1.3
blindspin==2.0.1
blivet==3.1.7
blivet-gui==2.1.15
Brlapi==0.7.0
cached-property==1.5.1
Cerberus==1.3.2
certifi==2018.10.15
cffi==1.12.3
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==7.1.1
colorama==0.4.1
cryptography==2.6.1
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.8
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
docker==4.0.2
docker-compose==1.25.4
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
evdev==1.1.2
filelock==3.0.12
first==2.0.1
fros==1.1
gpg==1.13.1
humanize==0.5.1
idna==2.8
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
iso8601==0.1.11
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.0
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterthemes==0.20.0
kiwisolver==1.1.0
langtable==0.0.52
lesscpy==0.14.0
libcomps==0.1.15
lutris==0.5.7.1
lxml==4.4.0
Mako==1.1.0.dev0
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
ntplib==0.3.3
numpy==1.18.2
olefile==0.46
ordered-set==3.1
packaging==20.1
pandocfilters==1.4.2
paramiko==2.6.0
parse==1.11.1
parso==0.6.2
Paste==3.2.2
pexpect==4.8.0
pickleshare==0.7.5
pid==2.2.3
Pillow==6.2.2
pipenv==2018.11.26
ply==3.11
productmd==1.27
prometheus-client==0.7.1
prompt-toolkit==3.0.4
ptyprocess==0.6.0
pwquality==1.4.2
pycairo==1.18.1
pycparser==2.14
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.43.0.2
pydbus==0.6.0
pyenchant==2.0.0
Pygments==2.6.1
PyGObject==3.34.0
pyinotify==0.9.6
pykickstart==3.21
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyparted==3.11.2
pyrsistent==0.15.7
PySocks==1.7.0
python-augeas==0.5.0
python-dateutil==2.8.1
python-dotenv==0.9.1
python-meh==0.48
pytz==2019.2
pyudev==0.21.0
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
requests==2.22.0
requests-file==1.4.3
requests-ftp==0.3.1
rpm==4.15.1
scour==0.37
selenium==3.141.0
selinux==2.9
semver==2.8.1
Send2Trash==1.5.0
sepolicy==1.1
setools==4.2.2
shellingham==1.2.7
simpleline==1.6
six==1.14.0
slip==0.6.4
slip.dbus==0.6.4
sos==3.8
soupsieve==1.9.2
systemd-python==234
Tempita==0.5.1
terminado==0.8.3
testpath==0.4.4
texttable==1.6.2
toml==0.10.0
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.7
vboxapi==1.0
virtualenv==20.0.13
virtualenv-clone==0.5.3
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.56.0
widgetsnbextension==3.5.1
yarg==0.1.9
zipp==3.1.0
